# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Sample API",
        "contact": {
            "name": "Erik Alexander, Friscilia Sutanto, Sri Wardani",
            "url": "https://uph.edu"
        },
        "version": "1.0"
    },
    "host": "webservice.app",
    "basePath": "",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/v1/khs": {
            "get": {
                "tags": [
                    "KHS"
                ],
                "summary": "Retrieves the collection of KHS resources.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "KHS collection.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/khs"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            },
            "post": {
                "tags": [
                    "KHS"
                ],
                "summary": "Data KHS",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Data KHS",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Users Token.",
                        "schema": {
                            "property": "token",
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
        "/api/v1/khs/{id}": {
            "get": {
                "tags": [
                    "KHS"
                ],
                "summary": "Retrieves the collection of KHS resources.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Jurusan collection.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/khs"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            },
            "put": {
                "tags": [
                    "KHS"
                ],
                "summary": "Data KHS",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "Data KHS",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Update.",
                        "schema": {
                            "property": "token",
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            },
            "delete": {
                "tags": [
                    "KHS"
                ],
                "summary": "Removes the KHS resource.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "KHS resource delete."
                    },
                    "401": {
                        "description": "Unauthorized Action."
                    },
                    "404": {
                        "description": "Resource not found."
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "khs": {
            "properties": {
                "nim": {
                    "type": "string"
                },
                "id_matkul": {
                    "type": "integer"
                },
                "absensi": {
                    "type": "string"
                },
                "tugas": {
                    "type": "string"
                },
                "uts": {
                    "type": "string"
                },
                "uas": {
                    "type": "string"
                },
                "grade": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "update_at": {
                    "type": "string"
                }
            }
        },
...
    }
}